Rename FrontendRouter source file and minor tweaks
