+++ config/initializers/frontend_router.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+require 'frontend_router'
