+++ spec/rails_helper.rb
@@ -38,6 +38,8 @@ RSpec.configure do |config|
 
   # Before the test suite is run
   config.before(:suite) do
+    UtmUrlBuilder.default_utm_source = nil
+
     begin
       # Since we're using Spring we must reload all factories
       # see: https://github.com/thoughtbot/factory_girl/blob/master/GETTING_STARTED.md#rails-preloaders-and-rspec
